refactor: Bench.sync => Bench.apps.sync
* Added missing imports
* Fixed broken import paths
