+++ bench/commands/update.py
@@ -44,5 +44,5 @@ def switch_to_branch(branch, apps, upgrade=False):
 @click.command('switch-to-develop')
 def switch_to_develop(upgrade=False):
 	"Switch frappe and erpnext to develop branch"
-	from bench.app import switch_to_develop
+	from bench.utils.app import switch_to_develop
 	switch_to_develop(apps=['frappe', 'erpnext'])
